Model the kernel in f64
Vec2/Vec3 are glam's DVec2/DVec3 and every coordinate, parameter and
tolerance in gridfinity-cad is f64. Four functions narrow to f32, all at
the very end and all because their consumer is: flat_vertices,
to_stl_binary (binary STL is f32 by spec), render_buffer and
welded_render_buffer (wgpu vertex buffers).

This retires the drawer's blend-chain kink. The refused chain put the
ball centre 4.236e-2 mm apart at one vertex, a 1.0114 deg turn where two
edges should be tangent. The configuration is a tangency -- a divider's
rounded end cap meets the neighbouring divider's face at exactly one
point -- and a tangency square-root-amplifies representation error, so a
9.35e-5 mm miss, about 12 f32 ulps at 81 mm, came out as 1 deg of turn
against the 0.5 deg MAX_JOIN_KINK allows. At f64 the same 12 ulps is
1.7e-13 mm and the turn is 4e-5 deg. The fitted drawer now lands 259 of
283 floor fillets where it landed 175 of 204.

It does not retire the fuzz backlog, which is structural: fuzz_inner_walls
21/150 -> 20/150 and fuzz_params_broad 7/400 -> 8/400. The large classes
are missing B-rep operators and no precision reaches them. It also costs
one profile -- fuzz_wall_openings green -> 2/150, a 0.175 mm sliver edge
left unpaired at a reentrant corner, undiagnosed and recorded as red.

Dir collapses to a newtype over Vec3: its exact/approx split existed only
because the kernel was f32, and dot_exact/cross_exact with it. The wasm
boundary widens to f64, which removes a narrowing rather than adding one
since a JavaScript number is already a double, and the packer's walls now
reach InnerWall with no cast at all. gridfinity-app stays mixed on
purpose, egui screen space being f32.

Tolerances that only ever absorbed f32 noise came down with it, with the
lib gate and all eight profiles byte-identical across the change:
END_AGREE 1e-4 -> 1e-9, ON_EDGE 1e-3 -> 1e-8, EXTREME_TIE_MM 1e-3 -> 1e-9,
SAME_PLANE_DIST 1e-4 -> 1e-9, SAME_PLANE_SIN 1e-5 -> 1e-10, PERP_TOL
1e-4 -> 1e-9, audit's GEO_TOL 1e-3 -> 1e-6. isect::TOL and the X_T
validator's ON_GEOMETRY_M are left alone: those stand for a feature size
and for the reader's own contract, not for the precision.

Cost: a fitted drawer builds in 0.67 s against 0.40 s and exports its
STLs in 1.05 s against 0.83 s.

Validation: cargo test --release --workspace green; all eight fuzz
profiles measured and recorded; optimize examples/drawer.toml green in
both formats; npm lint/test/build and build:wasm green.

+++ crates/gridfinity-app/src/editor.rs
@@ -30,8 +30,8 @@ pub struct Editor {
     pub active_bin: usize,
     drag_start: Option<Pos2>,
     drag_now: Option<Pos2>,
-    pub wall_width: f32,
-    pub wall_height: f32,
+    pub wall_width: f64,
+    pub wall_height: f64,
     pub wall_full: bool,
 }
 
@@ -147,14 +147,14 @@ impl Editor {
 
         for w in &p.inner_walls {
             let a = Pos2::new(
-                rect.left() + w.x1 / PITCH * cell,
-                rect.bottom() - w.y1 / PITCH * cell,
+                rect.left() + w.x1 as f32 / PITCH * cell,
+                rect.bottom() - w.y1 as f32 / PITCH * cell,
             );
             let b2 = Pos2::new(
-                rect.left() + w.x2 / PITCH * cell,
-                rect.bottom() - w.y2 / PITCH * cell,
+                rect.left() + w.x2 as f32 / PITCH * cell,
+                rect.bottom() - w.y2 as f32 / PITCH * cell,
             );
-            let px = (w.width.max(0.4) / PITCH * cell).max(2.0);
+            let px = (w.width.max(0.4) as f32 / PITCH * cell).max(2.0);
             painter.line_segment([a, b2], Stroke::new(px, Color32::from_rgb(0x8a, 0x5a, 0x2b)));
         }
         if let (Some(a), Some(b2)) = (self.drag_start, self.drag_now) {
@@ -199,10 +199,10 @@ impl Editor {
                         let (ax, ay) = to_grid(a);
                         let (bx, by) = to_grid(b2);
                         let w = InnerWall {
-                            x1: ax * PITCH,
-                            y1: ay * PITCH,
-                            x2: bx * PITCH,
-                            y2: by * PITCH,
+                            x1: (ax * PITCH) as f64,
+                            y1: (ay * PITCH) as f64,
+                            x2: (bx * PITCH) as f64,
+                            y2: (by * PITCH) as f64,
                             width: self.wall_width,
                             height: (!self.wall_full).then_some(self.wall_height),
                         };
